Fix focus bug in navbar
